obs-script: Add Lua/Python scripting plugin
Built upon the obs-python plugin originally made by Andyroid, it has
been greatly refactored, and extended to also support Lua in addition to
Python.

Lua is recommended for back-end scripts due to its performance and
ability to have separate script states for different threads, and Python
is recommended for front-end scripts and things unrelated to the
back-end.